Make the tool more reliable for our use-case by removing the cargo aspect of it
