+++ codegen-core/src/main/kotlin/software/amazon/smithy/rust/codegen/core/smithy/protocols/serialize/EventStreamErrorMarshallerGenerator.kt
@@ -97,7 +97,7 @@ class EventStreamErrorMarshallerGenerator(
                 rust("let mut headers = Vec::new();")
                 addStringHeader(":message-type", """"exception".into()""")
                 if (errorsShape.errorMembers.isEmpty()) {
-                    rust("let payload = Vec::new();")
+                    rust("let payload = Vec::new()")
                 } else {
                     rustBlock("let payload = match _input") {
                         errorsShape.errorMembers.forEach { error ->
